% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scale_reduction_code.R
\name{ScaleReduce}
\alias{ScaleReduce}
\title{Function to Analyze Survey Scales for Possible Reduction}
\usage{
ScaleReduce(full.scale, retest = FALSE, factors = 2)
}
\arguments{
\item{full.scale}{A dataset containing all items in the full scale, rescaled to range between 0 and 1; must not include NAs}

\item{retest}{Indicator for whether the scale was administered twice; in this case, responses from both administrations should be included in full.scale; default is FALSE}

\item{factors}{the number of factors you want to analyze in the principal components section; default is 2}
}
\value{
abbreviated.scales creates mean values of all subscales (including one-item scales and the full original scale)

abbreviated.scales.retest if retest=TRUE, creates mean values of all subscales from second administration of the scale

statistics provides information on each subscale, including which items the scale contains, its length, and all possible statistics calculated using this function (e.g., Cronbach's alpha cannot be calculated for the one-item scales, so these statistics are not provided here)

item.analysis summarizes information on subscales of various length (see tables X and Y in Loose, et. al. (2015))

question.analysis summarizes information on subscales containing each item from the original full scale (see tables X and Y in Loose, et. al. (2015))
}
\description{
This function provides a series of statistics to determine minimum scale length with properties similar to a full scale.
}
\author{
Krista Loose \email{loosek@mit.edu}
}
\references{
Loose, Krista, Yue Hou, Adam Berinsky, and Cindy Kam.  2015.  "Strategies for Scale Reduction."  Working Paper.
}

